+ New option to define delay time for outbound messages + New option to send note on for note off messages + Added support for include files in aircraft profiles

+++ src/profile/map/map_out/map_out_list.cpp
@@ -71,7 +71,7 @@ map_out_itr map_out_list::end()
 /**
  * Return the number of mapping
  */
-unsigned int map_out_list::size()
+size_t map_out_list::size()
 {
     return m_list.size();
 }
